appdev_genai_googlecloud/src/genai-app-firestore/requirements.txt
Flask==3.0.3
gunicorn==23.0.0
cloudevents==1.11.0
google-cloud-firestore==2.18.0
google-cloud-aiplatform==1.65.0
google-cloud-storage==2.18.2
tenacity==8.3.0
langchain-community==0.2.16
langchain-google-community==1.0.8
langchain-text-splitters==0.2.4
pypdf==4.3.1
nanoid==2.0.0